+++ src/main/scala/com/github/gchudnov/mtg/Diagram.scala
@@ -58,7 +58,7 @@ object Diagram:
     val (viewTicks, viewLabels) = if view.isLimited then
       val vi = view.toInterval
       val vs = toSpan(translator, vi)
-      (Span.toTicks(vs), canvas.labels(vi, vs))
+      (Span.toTicks(vs), Label.make(canvas, vi, vs))
     else (List.empty[Tick], List.empty[Label])
 
     val d = intervals.zipWithIndex.foldLeft(Diagram.empty) { case (acc, (i, j)) =>
@@ -66,7 +66,7 @@ object Diagram:
 
       val span   = toSpan(translator, i)
       val ticks  = Span.toTicks(span)
-      val labels = canvas.labels(i, span)
+      val labels = Label.make(canvas, i, span)
       val legend = Legend.make(i)
       val ann    = if j < annotations.size then Annotation(annotations(j)) else Annotation.empty
