+++ src/main/scala/com/github/gchudnov/mtg/diagram/Canvas.scala
@@ -1,11 +1,6 @@
 package com.github.gchudnov.mtg.diagram
 
-import com.github.gchudnov.mtg.Diagram.Label
-import com.github.gchudnov.mtg.Domain
-import com.github.gchudnov.mtg.Interval
 import com.github.gchudnov.mtg.Value
-import com.github.gchudnov.mtg.Mark
-import com.github.gchudnov.mtg.Show
 
 /**
  * Canvas
@@ -26,49 +21,7 @@ final case class Canvas(
 
   println(("left, right, first, last, size", left, right, first, last, size))
 
-  /**
-   * Make a label so that it is visible on the canvas.
-   */
-  def label(x: Int, text: String): Label =
-    val p = Canvas.align(x.toDouble - (text.size.toDouble / 2.0))
-    val q = p + text.size
-
-    // if label just slightly out of the bounds, make it visible
-    val x1 =
-      if p < 0 && isIn(x) then 0
-      else if q >= width && isIn(x) then width - text.size
-      else p
-
-    Label(x1, text)
-
-  def labels[T: Domain](i: Interval[T], span: Span)(using Ordering[Mark[T]]): List[Label] =
-    if i.isEmpty then List.empty[Label]
-    else if i.isPoint then List(label(span.x0, Show.str(i.left.eval)))
-    else
-      List(
-        label(span.x0, toLeftLabel(i.left)),
-        label(span.x1, toRightLabel(i.right))
-      )
-
-  private def toLeftLabel[T: Domain](left: Mark[T]): String =
-    left match
-      case Mark.At(x) =>
-        Show.str(x)
-      case Mark.Succ(xx) =>
-        Show.str(xx.eval)
-      case xx @ Mark.Pred(_) =>
-        Show.str(xx.eval)
-
-  private def toRightLabel[T: Domain](right: Mark[T]): String =
-    right match
-      case Mark.At(y) =>
-        Show.str(y)
-      case yy @ Mark.Succ(_) =>
-        Show.str(yy.eval)
-      case Mark.Pred(yy) =>
-        Show.str(yy.eval)
-
-  private def isIn(x: Int): Boolean =
+  def isIn(x: Int): Boolean =
     (x >= 0 && x < width)
 
 object Canvas:
